Update Sonoff-4CH-Pro.md
I have owned ch4 for over 2 years. Immediately. I immediately found random activations of the buttons and I tried everything (butondebounce, MQTT retain, etc.).
I have other ESP8266 devices with Tasmota firmware and nobody was giving me this problem.
Now I found that the problems did not depend on the Tasmota firmware or the MQTT but on the RF 433 module. This communicates directly with the stm32f030c6 MCU and simulates the activation of a button. I tested this solution for a week and it solved the problem.
There is some information online about this solution but not easily identifiable because those who find it attribute it to problems in the buttons, problems with MQTT, problems with tasmota.
I think it can be very useful for those who read this document to program the module with tasmota to be notified in advance so that we can solve it quickly.

I made an image with the description of the changes, how can I send it?